Support in control name backward compatibility
