+++ gitregostore/gitstoremethods.go
@@ -80,7 +80,7 @@ func (gs *GitRegoStore) GetOPAControlByName(controlName string) (*opapolicy.Cont
 	gs.controlsLock.RLock()
 	defer gs.controlsLock.RUnlock()
 	for _, control := range gs.Controls {
-		if strings.EqualFold(control.Name, controlName) {
+		if strings.EqualFold(control.Name, controlName) || strings.EqualFold(control.Name, getNewControlName(control.ControlID, controlName)) {
 			err := gs.fillRulesAndRulesIDsInControl(&control)
 			if err != nil {
 				return nil, err
